requirements.txt
torchvision~=0.21.0+cu126
torch~=2.6.0+cu126
pandas~=2.2.3
numpy~=2.1.1
pillow~=11.0.0